Add .travis.yaml
Because my PC is low performance, I use TravisCI to build this.
In addition, I would like to keep an error message when building.
I think that it will expire if I do not take a special way when
building this, but I think I'll start with the simplest setting.